requirements.txt
asgiref==3.8.1
coverage==7.6.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
drf-yasg==1.21.8
celery==5.3.6
redis==5.2.1
stripe==5.0.0
gunicorn==20.1.0
